+++ oap-server/analyzer/event-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/event/listener/EventRecordAnalyzerListener.java
@@ -22,6 +22,7 @@ import com.google.gson.Gson;
 import lombok.RequiredArgsConstructor;
 import org.apache.skywalking.apm.network.event.v3.Source;
 import org.apache.skywalking.oap.server.core.CoreModule;
+import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
 import org.apache.skywalking.oap.server.core.analysis.worker.MetricsStreamProcessor;
 import org.apache.skywalking.oap.server.core.config.NamingControl;
@@ -50,6 +51,8 @@ public class EventRecordAnalyzerListener implements EventAnalyzerListener {
 
     @Override
     public void parse(final org.apache.skywalking.apm.network.event.v3.Event e) {
+        event.setLayer(Layer.nameOf(e.getLayer()));
+
         event.setUuid(e.getUuid());
 
         if (e.hasSource()) {