+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/query/type/event/Event.java
@@ -24,7 +24,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.Data;
-import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.core.query.type.KeyValue;
 
@@ -48,7 +47,7 @@ public class Event {
 
     private long endTime;
 
-    private String layer = Layer.UNDEFINED.name();
+    private String layer;
 
     public void setParameters(final List<KeyValue> parameters) {
         this.parameters = parameters;